change to float64

+++ internal/sensors/bmp280.go
@@ -2,6 +2,6 @@ package sensors
 
 type BMP280Sensor struct {
 	Mac         string  `json:"mac"`
-	Pressure    float32 `json:"pres"`
-	Temperature float32 `json:"temp"`
+	Pressure    float64 `json:"pres"`
+	Temperature float64 `json:"temp"`
 }
