+++ internal/sensors/voltage.go
@@ -2,5 +2,5 @@ package sensors
 
 type VoltageSensor struct {
 	Mac     string  `json:"mac"`
-	Voltage float32 `json:"voltage"`
+	Voltage float64 `json:"voltage"`
 }
